refactor: 컴포넌트 내에 스타일 외부 파일로 이동
