+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import theme from '../../assets/styles/theme';
+import * as style from './FooterStyle';
+
+function Footer() {
+	return (
+		<style.Foot theme={theme}>© 2021 Friends - by Changyeop Song</style.Foot>
+	);
+}
+
+export default Footer;
