+++ src/components/Footer/FooterStyle.ts
@@ -1,14 +1,6 @@
-import React from 'react';
 import styled from 'styled-components';
-import theme from '../assets/styles/theme';
 
-function Footer() {
-	return <Foot theme={theme}>© 2021 Friends - by Changyeop Song</Foot>;
-}
-
-export default Footer;
-
-const Foot = styled.footer`
+export const Foot = styled.footer`
 	width: 100%;
 	height: 5vh;
 	display: flex;
