+++ src/page/DetailPage.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components';
 import theme from '../assets/styles/theme';
-import Check from '../components/Check';
-import Location from '../components/Location';
-import ModalFooter from '../components/ModalFooter';
+import Check from '../components/Check/Check';
+import Location from '../components/Location/Location';
+import ModalFooter from '../components/Modal/ModalFooter/ModalFooter';
 import NavContainer from '../containers/NavContainer';
 import { pageAction } from '../modules/CurrentPage/PageCheck';
 import { ReducerType } from '../modules/rootReducer';
